JOSS Code Review
## Alex changes for JOSS review

o For `cpAlgorithm`, `cpPermuteEntropy`, and `cpThreshold`, I made it so that any matrix can be input (rather than just qgraph objects). The code still does not specify for a square or symmetric matrix, which would be recommended. Having input as either a matrix (data.frame) or qgraph object allows your users greater flexibility

o Specific for `cpPermuteEntropy`: I added code to parallelize the function to increase the computational speed. Although the old code worked well for a small number of variables, I performed the analysis on one of my own datasets with 48 variables and this took 82.16 seconds. With the parallelization, the same analysis takes 22.25 (or about a quarter of the time). The parallelization should scale with the sample size -- the larger the sample, the greater the time efficiency

o Note that I changed the DESCRIPTION file to add the *parallel* package as an IMPORT

o I also changed the .Rd of `cpPermuteEntropy` to add the `ncores` argument. The default is for half the processing cores on the users computer